requirements.txt
nltk==3.7
mypy==0.961
PyYAML==6.0
praw==7.6.0
spacy==1.8.1
numpy==1.23.1
pandas==1.3.4
yaspin==2.2.0
gensim==4.2.0
progress==1.6
seaborn==0.11.2
pyldavis==3.3.1
requests==2.26.0
WordCloud==1.8.1
matplotlib==3.5.2
statsmodels==0.13.2
scikit-learn==1.1.1
pybliometrics==3.3.0
python-dotenv==0.20.0
mypy-extensions==0.4.3